requirements.txt
Django
django-import-export
django-rest-multiple-models
djangorestframework
views
gunicorn
django-heroku